Tidy up search screen - break search screen up
